Reuse not found page for 403 Forbidden errors
This provides a better user experience, as attempts to access private
resources now results in a more user-friendly error message.
Ideally the 403 status would be rewritten to a 404 Not Found status,
but the easiest solution I've found requires PHP or other similar
server-side scripting and I don't want to add that as a dependency.